exception handling

+++ src/main/java/com/example/servicemanagement/services/SProviderService.java
@@ -2,6 +2,8 @@ package com.example.servicemanagement.services;
 
 import com.example.servicemanagement.dtos.RequestSPDto;
 import com.example.servicemanagement.dtos.ResponseSPDto;
+import com.example.servicemanagement.exceptions.NotFoundException;
+import com.example.servicemanagement.exceptions.ValueNotAllowedException;
 import com.example.servicemanagement.models.ServiceProvider;
 import com.example.servicemanagement.repositories.SProviderRepository;
 import org.springframework.stereotype.Service;
@@ -17,8 +19,10 @@ public class SProviderService {
         this.sProviderRepository = sProviderRepository;
     }
 
-    public ResponseSPDto createServiceProvider(RequestSPDto requestDto) {
-        // todo: if the SP is already present
+    public ResponseSPDto createServiceProvider(RequestSPDto requestDto) throws Exception {
+        Optional<ServiceProvider> existingServiceProvider = sProviderRepository.findByName(requestDto.getName());
+        if(existingServiceProvider.isPresent())
+            throw new ValueNotAllowedException("VNA_001", "Service Provider with provided name is already present");
 
         ServiceProvider serviceProvider = new ServiceProvider();
         serviceProvider.setName(requestDto.getName());
@@ -33,7 +37,7 @@ public class SProviderService {
     public ResponseSPDto getServiceProvider(String id) throws Exception {
         Optional<ServiceProvider> serviceProvider = sProviderRepository.findById(id);
         if(serviceProvider.isEmpty())
-            throw new Exception("Service Provider is not present");
+            throw new NotFoundException("NF_001", "Service Provider");
         return ResponseSPDto.from(serviceProvider.get());
     }
 
@@ -47,7 +51,7 @@ public class SProviderService {
     public ResponseSPDto updateServiceProvider(RequestSPDto updateRequestDto, String id) throws Exception {
         Optional<ServiceProvider> optionalServiceProvider = sProviderRepository.findById(id);
         if(optionalServiceProvider.isEmpty())
-            throw new Exception("Service Provider is not present");
+            throw new NotFoundException("NF_002", "Service Provider");
 
         ServiceProvider existingServiceProvider = optionalServiceProvider.get();
         existingServiceProvider.setName(updateRequestDto.getName());
@@ -59,7 +63,11 @@ public class SProviderService {
         return ResponseSPDto.from(updatedServiceProvider);
     }
 
-    public ResponseSPDto deleteServiceProvider(String id) {
+    public ResponseSPDto deleteServiceProvider(String id) throws NotFoundException {
+        Optional<ServiceProvider> optionalServiceProvider = sProviderRepository.findById(id);
+        if(optionalServiceProvider.isEmpty())
+            throw new NotFoundException("NF_003", "Service Provider");
+
         sProviderRepository.deleteById(id);
         return ResponseSPDto.empty();
     }
